Enum numbas::question::answer_simplification::AnswerSimplificationRule

pub enum AnswerSimplificationRule {

    All(bool),
    Basic(bool),
    CancelUnitFactors(bool),
    CancelUnitPowers(bool),
    CancelUnitDenominators(bool),
    CancelZeroFactors(bool),
    OmitZeroTerms(bool),
    CancelZeroPowers(bool),
    NoLeadingMinus(bool),
    CollectNumbers(bool),
    Fractions(bool),
    CancelPowersWithBaseZero(bool),
    ConstantsFirst(bool),
    CollectSqrtProducts(bool),
    CollectSqrtDivisions(bool),
    CancelSqrtSquares(bool),
    Trigonometric(bool),
    EvaluatePowersOfNumbers(bool),
    CollectTerms(bool),
    CollectPowersOfCommonFactors(bool),
    CollectLikeFractions(bool),
    CanonicalOrder(bool),
    ExpandBrackets(bool),

}

See https://docs.numbas.org.uk/en/latest/simplification.html#id1 for examples

Variants

All(bool)

To turn all built-in rules on, use the name all. To turn all built-in rules off, use !all. Note: Because they can conflict with other rules, the canonicalOrder and expandBrackets rules are not included in all. You must include them separately.

Basic(bool)

Basic simplifications like unary plus removal etc

CancelUnitFactors(bool)

Cancel products of 1

CancelUnitPowers(bool)

Cancel exponents of 1

CancelUnitDenominators(bool)

Cancel fractions with denominator 1

CancelZeroFactors(bool)

Cancel products of zero to zero

OmitZeroTerms(bool)

Omit zero terms

CancelZeroPowers(bool)

Cancel exponents of 0

NoLeadingMinus(bool)

Rearrange expressions so they don’t start with a unary minus

CollectNumbers(bool)

Collect together numerical (as opposed to variable) products and sums.

Fractions(bool)

Cancel fractions to lowest form.

CancelPowersWithBaseZero(bool)

Cancel any power of zero

ConstantsFirst(bool)

Numbers go to the left of multiplications

CollectSqrtProducts(bool)

Collect products of square roots

CollectSqrtDivisions(bool)

Collect fractions of square roots

CancelSqrtSquares(bool)

Cancel square roots of squares, and squares of square roots

Trigonometric(bool)

Simplify some trigonometric identities

EvaluatePowersOfNumbers(bool)

Evaluate powers of numbers.

CollectTerms(bool)

Collect together and cancel terms. Like collectNumbers, but for any kind of term.

CollectPowersOfCommonFactors(bool)

Collect together powers of common factors.

CollectLikeFractions(bool)

Collect together fractions over the same denominator.

CanonicalOrder(bool)

earrange the expression into a “canonical” order, using canonical_compare(). Note: This rule can not be used at the same time as noLeadingMinus - it can lead to an infinite loop.

ExpandBrackets(bool)

Expand out products of sums. (distributivity)