Struct numbas::question::Question

pub struct Question {

    pub name: String,
    pub statement: ContentAreaString,
    pub advice: ContentAreaString,
    pub parts: Vec<QuestionPart>,
    pub builtin_constants: BuiltinConstants,
    pub constants: Vec<QuestionConstant>,
    pub variables: BTreeMap<String, Variable>,
    pub variables_test: QuestionVariablesTest,
    pub functions: BTreeMap<String, Function>,
    pub ungrouped_variables: Vec<String>,
    pub variable_groups: Vec<VariableGroup>,
    pub rulesets: BTreeMap<String, Vec<AnswerSimplificationType>>,
    pub preamble: Preamble,
    pub navigation: QuestionNavigation,
    pub extensions: Vec<String>,
    pub resources: Vec<Resource>,
    pub tags: Vec<String>,
    pub custom_part_types: Vec<CustomPartType>,

}

Fields
name: Stringstatement: ContentAreaString

The statement is a content area which appears at the top of the question, before any input boxes. Use the statement to set up the question and provide any information the student needs to answer it.

advice: ContentAreaString

Advice is a content area which is shown when the student presses the Reveal button to reveal the question’s answers, or at the end of the exam. The advice area is normally used to present a worked solution to the question.

parts: Vec<QuestionPart>builtin_constants: BuiltinConstantsconstants: Vec<QuestionConstant>variables: BTreeMap<String, Variable>variables_test: QuestionVariablesTestfunctions: BTreeMap<String, Function>ungrouped_variables: Vec<String>variable_groups: Vec<VariableGroup>rulesets: BTreeMap<String, Vec<AnswerSimplificationType>>preamble: Preamblenavigation: QuestionNavigationextensions: Vec<String>resources: Vec<Resource>tags: Vec<String>

Tags starting with ’skill: ’ are used in diagnostic mode to specify a topic

custom_part_types: Vec<CustomPartType>
Implementations

impl Question

pub fn from_question_exam_str(s: &str) -> Result<Question>