Module numbas::jme

Modules

ast

builtin_functions

parser

Structs

ContentAreaString

Each portion of text displayed to the student (for example, the statement, advice, and part prompts) is a content area. A content area can include text, images, or more dynamic content such as videos and interactive diagrams.

ContentAreaStringChange

ContentAreaStringDesc

EmbracedJMEString

EmbracedJMEStringChange

EmbracedJMEStringDesc

JMENotesString

Each portion of text displayed to the student (for example, the statement, advice, and part prompts) is a content area. A content area can include text, images, or more dynamic content such as videos and interactive diagrams.

JMENotesStringChange

JMENotesStringDesc

JMEString

JMEStringChange

JMEStringDesc